Reuse IP range validator
